requirements.txt
Flask==2.3.3
flask-cors==4.0.0
PyMySQL==1.1.0
python-dotenv==1.0.0
gunicorn==21.2.0